requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==22.2.0
autopep8==2.3.2
bcrypt==4.3.0
build==0.10.0
CacheControl==0.12.11
cachetools==5.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
click==8.2.0
crashtest==0.4.1
cryptography==39.0.2
distlib==0.3.6
dnspython==2.7.0
dulwich==0.21.3
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
filelock==3.9.0
flake8==7.2.0
google-api-core==2.24.2
google-auth==2.40.1
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.1.0
installer==0.6.0
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.17.3
jwt==1.3.1
keyring==23.13.1
lockfile==0.12.2
mccabe==0.7.0
more-itertools==9.1.0
msgpack==1.0.5
packaging==23.0
passlib==1.7.4
pexpect==4.8.0
pipenv==2023.2.18
pkginfo==1.9.6
platformdirs==2.6.2
pluggy==1.5.0
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
proto-plus==1.26.1
protobuf==5.29.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.21
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyflakes==3.3.2
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
python-jose==3.4.0
PyYAML==6.0.2
rapidfuzz==2.15.1
redis==6.1.0
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9.1
SecretStorage==3.3.3
setuptools==80.9.0
shellingham==1.5.0.post1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tomlkit==0.11.6
trove-classifiers==2023.3.9
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==1.26.15
uvicorn==0.34.2
uvloop==0.21.0
virtualenv==20.20.0
virtualenv-clone==0.5.7
watchfiles==1.0.5
webencodings==0.5.1
websockets==15.0.1
zipp==3.15.0
